ui break down into component
