+++ app/index.tsx
@@ -1,19 +1,8 @@
 import ScreenTwo from "@/component/ScreenTwo";
-import { Button, StyleSheet, Text, TouchableOpacity, View } from "react-native";
+import { View } from "react-native";
 
 
 export default function Index() {
-
-  const clickbutton = () => {
-    console.log("button clicked")
-  }
-
-  /* for touchableopacity */
-  const onPress = () => {
-    console.log("opacity button pressed")
-  }
-
-
 
   return (
     <View
@@ -24,33 +13,14 @@ export default function Index() {
         backgroundColor: "hsl(200, 100%, 50%)",
       }}
     >
-      <Text style={styles.text}>one ticket</Text>
-      <Text >global ticket</Text>
-      <Button title="Click me!" onPress={() => alert("Hello!")} />
-      <Button title="press me!" onPress={clickbutton} />
-      <TouchableOpacity activeOpacity={0.6} style={styles.button} onPress={onPress}>
-          <Text>Press Here</Text>
-      </TouchableOpacity>
-
       <ScreenTwo />
-
-      
     </View>
   );
 }
 
-const styles = StyleSheet.create({
-  text: {
-    fontWeight: "bold",
-    fontSize: 30,
-    textTransform: "uppercase",
-    color: "white",
-  },
 
-  button: {
-    marginTop: 20,
-    backgroundColor: 'blue',
-    padding: 10,
-    borderRadius: 5,
-  },
-});
+
+
+  
+
+
